@bobg 2017-05-19
PTAL
Backporting https://github.com/chain/chain/pull/1216 to the 1.2-stable branch